requirements.txt
appnope>=0.1.0
Babel>=2.5.1
backports.shutil-get-terminal-size>=1.0.0
construct>=2.5.3
decorator>=4.1.2
Django>=1.11.6
django-ckeditor>=5.3.1
django-el-pagination>=3.1.0
django-formtools>=2.1
django-haystack>=2.6.1
django-js-asset>=0.1.1
django-otp>=0.4.1.1
django-phonenumber-field>=1.3.0
django-two-factor-auth>=1.6.2
django-user-sessions>=1.5.3
django-simple-captcha>=0.5.5
django-activelink>=1.0
enum34>=1.1.6
future>=0.16.0
ipdb>=0.10.3
ipython>=5.5.0
ipython-genutils>=0.2.0
pathlib2>=2.3.0
pefile>=2017.9.3
pexpect>=4.2.1
phonenumberslite>=8.8.4
pickleshare>=0.7.4
prompt-toolkit>=1.0.15
ptyprocess>=0.5.2
Pygments>=2.2.0
python-ptrace>=0.9.3
pytz>=2017.3
qrcode>=4.0.4
scandir>=1.6
simplegeneric>=0.8.1
six>=1.11.0
traitlets>=4.3.2
wcwidth>=0.1.7
Whoosh>=2.7.4
geoip2>=2.5.0
